requirements.txt
asgiref==3.5.0
astroid==2.9.0
attrs==19.3.0
Automat==0.8.0
backports.zoneinfo==0.2.1
black==21.11b1
bleach==4.1.0
blinker==1.4
cached-property==1.5.1
certifi==2019.11.28
chardet==3.0.4
click==8.0.3
cloud-init==21.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.3
django-active-link==0.1.8
django-cors-headers==3.11.0
django-summernote==0.8.20.0
django-taggit==2.0.0
djangorestframework==3.13.1
docker==4.1.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
filelock==3.4.2
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
isort==5.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==4.2.0
mypy-extensions==0.4.3
netifaces==0.10.4
numpy==1.22.2
oauthlib==3.1.0
packaging==21.3
pandas==1.4.1
pathspec==0.9.0
pbr==5.8.0
pexpect==4.6.0
platformdirs==2.4.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.12.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.6
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2021.3
PyYAML==5.3.1
regex==2021.11.10
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
sos==4.2
sqlparse==0.4.2
ssh-import-id==5.10
stevedore==3.5.0
systemd-python==234
texttable==1.6.2
toml==0.10.2
tomli==1.2.2
Twisted==18.9.0
typing-extensions==4.0.1
ubuntu-advantage-tools==27.6
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.13.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wadllib==1.3.3
webencodings==0.5.1
websocket-client==0.53.0
wrapt==1.13.3
zipp==1.0.0
zope.interface==4.7.1
